# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Web App API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [rag]
      summary: List all uploaded documents
      description: Returns all documents in the knowledge base
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [rag]
      summary: Upload a document
      description: Uploads a document for indexing and RAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [rag]
      summary: Get document details
      description: Returns a single document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteDocument
      tags: [rag]
      summary: Delete a document
      description: Removes a document from the knowledge base
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /documents/search:
    post:
      operationId: searchDocuments
      tags: [rag]
      summary: Search knowledge base
      description: Searches documents using semantic or keyword search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchQuery"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        fileType:
          type: string
        size:
          type: integer
        createdAt:
          type: string
          format: date-time
        indexed:
          type: boolean
        chunkCount:
          type: integer
      required:
        - id
        - title
        - createdAt
        - indexed
        - chunkCount
    DocumentInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        fileType:
          type: string
      required:
        - title
        - content
    SearchQuery:
      type: object
      properties:
        query:
          type: string
        topK:
          type: integer
        searchType:
          type: string
          enum: [semantic, keyword, hybrid]
      required:
        - query
    SearchResult:
      type: object
      properties:
        documentId:
          type: integer
        title:
          type: string
        content:
          type: string
        score:
          type: number
        rank:
          type: integer
      required:
        - documentId
        - title
        - content
        - score
        - rank
...